Reset quantity and save database
-creation of menu to show action bar with one item delete
- action bar (toolbar) with delete button to reset all quantities
selected in the recycler view list
- save and restore instance state with information of current quantities
 selected by user, to not lose information if activity is stopped
- if activity is the first time to be executed delete all rows in the
table of the database, if activity restart from previous execution just
open the database
- close database connection to free resources while this activity is not
 in execution

+++ app/src/main/java/com/gmail/davidecoffaro/productscity/BuyProductsActivity.java
@@ -1,13 +1,18 @@
 package com.gmail.davidecoffaro.productscity;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 import androidx.room.Room;
 
 import android.content.Intent;
 import android.os.Bundle;
+import android.view.MenuItem;
 import android.view.View;
+import android.view.Menu;
 import android.widget.Button;
 import android.widget.TextView;
 import android.widget.Toast;
@@ -17,7 +22,6 @@ import com.gmail.davidecoffaro.productscity.utilclass.Prodotto;
 import com.gmail.davidecoffaro.productscity.utilclass.RVBuyProductListAdapter;
 import com.gmail.davidecoffaro.productscity.utilclass.databaseclass.NegozioDatabase;
 import com.gmail.davidecoffaro.productscity.utilclass.task.DatabaseTask;
-import com.gmail.davidecoffaro.productscity.utilclass.task.InsertAllDatabaseTask;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -29,12 +33,17 @@ public class BuyProductsActivity extends AppCompatActivity implements View.OnCli
     Button next;
     DataNegozioJSon negozioJSon;
     NegozioDatabase db;
+    Toolbar toolbar;
 
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_buy_products);
+
+        //setting up toolbar
+        toolbar = (Toolbar) findViewById(R.id.toolbarBuyProductList);
+        setSupportActionBar(toolbar);
 
         totalOrder = (TextView) findViewById(R.id.textViewTotalOrder);
         next = (Button) findViewById(R.id.buttonNext);
@@ -50,14 +59,73 @@ public class BuyProductsActivity extends AppCompatActivity implements View.OnCli
         db = Room.databaseBuilder(getApplicationContext(),
                 NegozioDatabase.class, "database-negozio1").addMigrations(NegozioDatabase.MIGRATION_1_2).build();
 
-        DatabaseTask deleteAll = new DatabaseTask(this, "DeleteAll");
-        deleteAll.execute(new Prodotto[0]);
+        //first execution of onCreate, reset database
+        if(savedInstanceState==null){
+            DatabaseTask deleteAll = new DatabaseTask(this, "DeleteAll");
+            deleteAll.execute(new Prodotto[0]);
+        }
 
-        //initialize recyclerView without setting the adapter, it will be setted when the asyncTask
+
+        //initialize recyclerView without setting the adapter, it will be set when the asyncTask
         // database finishes the query execution
         rv = (RecyclerView) findViewById(R.id.recyclerViewBuyProductList);
         rv.setLayoutManager(new LinearLayoutManager(this));
 
+    }
+
+    @Override
+    protected void onRestart() {
+        super.onRestart();
+
+        //creation of database with table Prodotto, using abstraction provided by Room Database
+        // table and rows of a database are linked to a java object
+        db = Room.databaseBuilder(getApplicationContext(),
+                NegozioDatabase.class, "database-negozio1").addMigrations(NegozioDatabase.MIGRATION_1_2).build();
+    }
+
+    @Override
+    protected void onRestoreInstanceState(@NonNull Bundle savedInstanceState) {
+        super.onRestoreInstanceState(savedInstanceState);
+
+        //restore information on the rows of the recyclerView with the values saved in the database
+        // onSaveInstanceState
+        getAllProducts();
+    }
+
+    @Override
+    protected void onSaveInstanceState(@NonNull Bundle outState) {
+        //save current quantities in the recycler view in the database to not lose this information
+        updateDatabase();
+
+        super.onSaveInstanceState(outState);
+    }
+
+    //action bar management
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        getMenuInflater().inflate(R.menu.menu_buyproductlist, menu);
+        return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(@NonNull MenuItem item) {
+        if(item.getItemId()==R.id.action_reset){
+            int i=0;
+            for(Prodotto p : negozioJSon.getListaprodotti()){
+                //reset all product's quantity to zero
+                p.setQuantita(0);
+                //notify adapter of change in its list
+                adapter.notifyDataSetChanged();
+                i++;
+            }
+
+            //reset total order
+            totalOrder.setText("0.0");
+
+            return true;
+        }else{
+            return  super.onOptionsItemSelected(item);
+        }
     }
 
     public void insertAllProducts(){
@@ -70,6 +138,7 @@ public class BuyProductsActivity extends AppCompatActivity implements View.OnCli
     }
 
     public void getAllProducts(){
+        //async task to get all products in the table Prodotto in database
         DatabaseTask getAll = new DatabaseTask(this, "GetAll");
         getAll.execute(new Prodotto[0]);
     }
@@ -96,28 +165,48 @@ public class BuyProductsActivity extends AppCompatActivity implements View.OnCli
             // una ArrayList<Prodotti> creati a partire dalla query di database, le due liste non
             // coincidono, quindi ho bisogno dell'update righe database
 
-            //if total order is 0, no products was buyed
-            if(totalOrder.getText().toString().equals("0.0")){
+            //if total order is 0 or less, no products was buyed
+            if(Float.parseFloat(totalOrder.getText().toString())<=0f){
                 Toast.makeText(this, R.string.error_no_product,Toast.LENGTH_SHORT).show();
             }else{
-                //take listaProdotti from negozioJSon and update table Prodotto in database with these
-                // new data (with quantity changed)
-                Prodotto[] paramProdottoDatabase = new Prodotto[negozioJSon.getListaprodotti().size()];
-                paramProdottoDatabase = negozioJSon.getListaprodotti().toArray(paramProdottoDatabase);
-
-                DatabaseTask update = new DatabaseTask(this, "Update");
-                update.execute(paramProdottoDatabase);
+                //save info quantities on database
+                updateAndStartCustomerCartActivityDatabase();
             }
-
-
-
         }
     }
 
+    private void updateDatabase(){
+        //take listaProdotti from negozioJSon and update table Prodotto in database with these
+        // new data (with quantity changed)
+        Prodotto[] paramProdottoDatabase = new Prodotto[negozioJSon.getListaprodotti().size()];
+        paramProdottoDatabase = negozioJSon.getListaprodotti().toArray(paramProdottoDatabase);
+
+        DatabaseTask update = new DatabaseTask(this, "Update");
+        update.execute(paramProdottoDatabase);
+    }
+
+    private void updateAndStartCustomerCartActivityDatabase(){
+        //take listaProdotti from negozioJSon and update table Prodotto in database with these
+        // new data (with quantity changed)
+        Prodotto[] paramProdottoDatabase = new Prodotto[negozioJSon.getListaprodotti().size()];
+        paramProdottoDatabase = negozioJSon.getListaprodotti().toArray(paramProdottoDatabase);
+
+        DatabaseTask update = new DatabaseTask(this, "UpdateStartNewActivity");
+        update.execute(paramProdottoDatabase);
+    }
+
     public void startCustomerCartActivity(){
+        //close database connection
+        closeDatabaseConnection();
+
         //start next activity
         Intent i = new Intent(this, CustomerCartActivity.class);
         i.putExtra("EXTRA_RIDER_MAIL", negozioJSon.getMailrider());
         startActivity(i);
+    }
+
+    public void closeDatabaseConnection(){
+        //close connection to database to free system resources
+        db.close();
     }
 }
